+++ CHANGES.rst
@@ -13,8 +13,10 @@ Unreleased
 * Increase timeout of the button actions from 2 to 10 seconds.
 * :meth:`~Reader.get_entries` now sorts entries by the import date first,
   and then by :attr:`~Entry.published`/:attr:`~Entry.updated`. (`#106`_)
+* Add ``enclosure_dedupe`` plugin (deduplicate enclosures of an entry). (`#78`_)
 
 .. _#106: https://github.com/lemon24/reader/issues/106
+.. _#78: https://github.com/lemon24/reader/issues/78
 
 
 Version 0.6